Add support for stream or normal
